+++ src/NWire.h
@@ -51,7 +51,7 @@ private:
 
 public:
     NWireHost();
-    uint8_t getData(uint8_t, uint8_t);
+    uint32_t getData(uint8_t, uint8_t);
     void sendData(uint8_t, uint8_t, uint32_t);
     uint8_t *getBuffer() const;
 };
